FIX: Fix architecture - rename ui module to ui-system, rename some files to match contents.
