+++ android/ui-system/src/main/AndroidManifest.xml
