+++ android/ui-system/src/main/res/anim/fade_in.xml
