+++ android/ui-system/src/main/res/anim/fade_out.xml
