+++ android/ui-system/src/main/res/anim/slide_out.xml
