+++ android/ui-system/src/main/res/drawable-anydpi/ic_arrow_back.xml
