+++ android/ui-system/src/main/res/values/strings.xml
