+++ settings.gradle.kts
@@ -37,7 +37,7 @@ enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
 include(":android:app")
 include(":android:library:architecture-test")
 include(":android:rocket")
-include(":android:ui")
+include(":android:ui-system")
 include(":shared:shared")
 include(":shared:common")
 include(":shared:rocket")
